requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==4.4.2
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
loguru==0.7.2
moviepy==1.0.3
numpy==1.26.4
pillow==10.2.0
proglog==0.1.10
pyaes==1.6.1
pyasn1==0.5.1
pydub==0.25.1
python-dotenv==1.0.1
requests==2.31.0
rsa==4.9
Telethon==1.34.0
tqdm==4.66.2
urllib3==2.2.1
